+++ src/main/java/edu/sdonohue/advancedjava/service/stocks/BasicStockService.java
@@ -1,6 +1,7 @@
-package edu.sdonohue.advancedjava.stocks;
+package edu.sdonohue.advancedjava.service.stocks;
 
 
+import edu.sdonohue.advancedjava.model.StockQuote;
 import org.jetbrains.annotations.Nullable;
 
 import javax.validation.constraints.NotNull;
