+++ src/main/java/edu/sdonohue/advancedjava/service/userstocks/UserStockService.java
@@ -1,4 +1,4 @@
-package edu.sdonohue.advancedjava.service;
+package edu.sdonohue.advancedjava.service.userstocks;
 
 import edu.sdonohue.advancedjava.model.Person;
