+++ src/main/java/edu/sdonohue/advancedjava/service/userstocks/UserStockServiceException.java
@@ -1,21 +1,9 @@
-package edu.sdonohue.advancedjava.stocks;
+package edu.sdonohue.advancedjava.service.userstocks;
 
 /**
- * Used to signal a problem with the StockService.
+ * Used to signal an issue with UserStockService
  */
-public class StockServiceException extends Exception {
-
-    /**
-     * Constructs a new exception with the specified detail message.  The
-     * cause is not initialized, and may subsequently be initialized by
-     * a call to {@link #initCause}.
-     *
-     * @param message the detail message. The detail message is saved for
-     *                later retrieval by the {@link #getMessage()} method.
-     */
-    public StockServiceException(String message) {
-        super(message);
-    }
+public class UserStockServiceException extends Exception {
 
     /**
      * Constructs a new exception with the specified detail message and
@@ -31,7 +19,7 @@ public class StockServiceException extends Exception {
      *                unknown.)
      * @since 1.4
      */
-    public StockServiceException(String message, Throwable cause) {
+    public UserStockServiceException(String message, Throwable cause) {
         super(message, cause);
     }
 }
