+++ src/test/java/edu/sdonohue/advancedjava/BasicStockServiceTest.java
@@ -1,8 +1,8 @@
 package edu.sdonohue.advancedjava;
 
-import edu.sdonohue.advancedjava.stocks.StockService.IntervalEnum;
-import edu.sdonohue.advancedjava.stocks.BasicStockService;
-import edu.sdonohue.advancedjava.stocks.StockQuote;
+import edu.sdonohue.advancedjava.service.stocks.StockService.IntervalEnum;
+import edu.sdonohue.advancedjava.service.stocks.BasicStockService;
+import edu.sdonohue.advancedjava.model.StockQuote;
 import org.junit.Before;
 import org.junit.Test;
